requirements.txt
python-dotenv
pytest
flask
requests
markdown-it-py>=2.2.0
markdown-it-py[plugins]
mdit-py-plugins
